fix(windows): one shared Authenticode reader for release tooling
The first pilot promotion failed at candidate verification with the exact
empty-status failure #195 fixed in the release validator:

  windows-release-candidate-verify: FAIL (Candidate Authenticode status
  must be Valid, got .)

windows-release-candidate-verify.mjs still carried its own copy of the old
powershell.exe reader — non-terminating errors, exit 0, discarded stderr.
Same bug, second copy.

Extract the proven pwsh-first reader from #195 into
scripts/lib/windows-authenticode.mjs and import it from both the release
validator and the candidate verifier, so the two cannot drift again.
